+++ test/methods/partition_test.js
@@ -6,8 +6,31 @@ module.exports = (it, expect, collect) => {
 
     const arr = collection.partition(i => i < 3);
 
-    expect(arr[0]).to.eql([1, 2]);
-    expect(arr[1]).to.eql([3, 4, 5, 6]);
+    expect(arr.all()).to.eql([
+      collect([1, 2]),
+      collect([3, 4, 5, 6]),
+    ]);
+    expect(arr.first().all()).to.eql([1, 2]);
     expect(collection.all()).to.eql([1, 2, 3, 4, 5, 6]);
+  });
+
+  it('should also work when collection is based on an object', () => {
+    const collection = collect({
+      name: 'Sadio Mané',
+      number: 19,
+      club: 'Liverpool FC',
+    });
+
+    const arr = collection.partition(v => typeof v === 'string');
+
+    expect(arr.all()).to.eql([
+      collect({
+        name: 'Sadio Mané',
+        club: 'Liverpool FC',
+      }),
+      collect({
+        number: 19,
+      }),
+    ]);
   });
 };